Allow option to use form data instead of json
